requirements.txt
requests
beautifulsoup4
pandas
lxml
openpyxl
python-dotenv